refactor(ci): remove internal maintainer tooling from public repo
Keep contributor-facing CI, failure comments, and CI_CHECKS.md. Remove
MAINTAINER.md, triage/review commands, and maintainer_review.py.
